+++ multiplication/multiplication.h
@@ -4,6 +4,6 @@
 #include <stdio.h>
 
 
-void mult(unsigned long long p1[], unsigned long long p2[], unsigned long long r[], size_t size, size_t r_size);
+void mult(unsigned long long p1[], unsigned long long p2[], unsigned long long r[], int size, int r_size);
 
 #endif // MULTIPLICATION
